feat: add AirPlay karaoke screensaver via shairport-sync
Extract shared karaoke rendering into KaraokeBase class and refactor
SonosKaraoke to extend it. Add AirPlayKaraoke screensaver that reads
shairport-sync's metadata pipe for track info and playback progress.
